netlabelctl: Add support for specifying address families.
The main user-visible change is to the map add command where
one can specify a '4' or a '6' as an optional protocol extra
value for the unlabelled protocol. eg:
  netlabelctl map add domain:foo_t protocol:unlbl,4
will map IPv4 addresses on domain foo_t to the unlabelled protocol.
Omitting the extra value will map all address families to that
protocol.

A similar notation is used when listing maps.  Domains displayed
as:
  domain:"foo_t",UNLABELED
have the mapping applied across all address families, while domains
displayed as:
  domain:"foo2_t",UNLABELED,4
would have the mapping for just IPv4.

Signed-off-by: Huw Davies <huw@codeweavers.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>